feat(auth): add authentication page component
